修改首页提示  修改退出亲友团

+++ src/sub-decorate/pages/friends/friends.vue
@@ -77,7 +77,9 @@
 					title: "是否确定退出亲友团？",
 					content: "退出后将无法看到亲友的房屋",
 					success: function(res) {
-						vm.deleteFriend(item.id);
+						if (res.confirm) {
+							vm.deleteFriend(item.id);
+						}
 					},
 				});
 			},
